objectStorageBackend/requirements.txt
boto3==1.34.136
botocore==1.34.136
Django==5.0.6
djangorestframework==3.15.2
python-decouple==3.8
django-cors-headers

pytest~=8.2.2